refactor: extract job status enum to types file
