+++ src/shared/jobs/__mocks__/enqueueJob.ts
@@ -1,6 +1,6 @@
 import { ulid } from "@/shared/ulid";
 import { beforeEach, vitest } from "vitest";
-import { JobStatus } from "../model";
+import { JobStatus } from "../types";
 
 export const enqueueJob = vitest.fn(
   (options: { type: string; payload?: any; parentJobId?: string }) => {
