+++ src/shared/jobs/types.ts
@@ -1,0 +1,6 @@
+export enum JobStatus {
+  Pending = "pending",
+  InProgress = "in-progress",
+  Complete = "complete",
+  Failed = "error",
+}
